requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0

# Data & Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
datasets>=2.14.0

# Evaluation Metrics
nltk>=3.8.0
rouge-score>=0.1.2
evaluate>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
